requirements.txt
absl-py==1.4.0
alabaster==0.7.13
amqp==5.1.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
astroid==2.15.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
billiard==3.6.4.0
binaryornot==0.4.4
black==23.3.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.2
crispy-bootstrap5==0.7
cron-descriptor==1.2.35
cryptography==39.0.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
dj-database-url==1.2.0
dj-rest-auth==3.0.0
Django==4.1.8
django-allauth==0.54.0
django-celery-beat==2.5.0
django-cors-headers==3.14.0
django-coverage-plugin==3.0.0
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==22.1
django-model-utils==4.3.1
django-modeltranslation==0.18.9
django-redis==5.2.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-stubs==1.16.0
django-stubs-ext==0.8.0
django-timezone-field==5.0
django-translations==1.3.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djangorestframework-stubs==1.10.0
docutils==0.19
drf-spectacular==0.26.1
drf-yasg==1.21.5
executing==1.2.0
factory-boy==3.2.1
Faker==18.4.0
filelock==3.9.0
flake8==6.0.0
flake8-isort==6.0.0
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
gunicorn==20.1.0
h5py==3.8.0
identify==2.5.22
idna==3.4
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
ipdb==0.13.13
ipython==8.12.0
isort==5.12.0
itypes==1.2.0
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.17.3
keras==2.12.0
kombu==5.2.4
lazy-object-proxy==1.9.0
libclang==16.0.0
livereload==2.6.3
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
ml-dtypes==0.1.0
mypy==1.1.1
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.0.0
pluggy==1.0.0
pre-commit==3.2.2
prompt-toolkit==3.0.38
protobuf==4.22.4
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==1.7.1
pylint==2.17.2
pylint-celery==0.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyrsistent==0.19.3
pytest==7.2.2
pytest-django==4.5.2
pytest-sugar==0.9.6
python-crontab==2.7.1
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
stack-data==0.6.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.11.7
tornado==6.2
traitlets==5.9.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.9
types-requests==2.28.11.17
types-urllib3==1.26.25.10
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
virtualenv==20.19.0
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
Werkzeug==2.2.3
whitenoise==6.4.0
wrapt==1.14.1
